+++ namer/k8s/src/main/scala/io/buoyant/namer/k8s/K8sInitializer.scala
@@ -27,7 +27,8 @@ object K8sInitializer extends K8sInitializer
 
 case class K8sConfig(
   host: Option[String],
-  port: Option[Port]
+  port: Option[Port],
+  labelSelector: Option[String]
 ) extends NamerConfig with ClientConfig {
 
   @JsonIgnore
@@ -49,6 +50,6 @@ case class K8sConfig(
       val label = param.Label(s"namer${prefix.show}/$ns")
       Api(client.configured(label).newService(dst)).withNamespace(ns)
     }
-    new EndpointsNamer(prefix, mkNs)
+    new EndpointsNamer(prefix, labelSelector, mkNs)
   }
 }